1. AI and ML Basics/10. LlamaIndex/Document-gpt-llamaindex/requirements.txt
streamlit
llama-index
faiss-cpu
sentence-transformers
PyMuPDF  # for PDF reading
docx2txt  # for DOCX
